+++ premium/grunt/config/browserify.js
@@ -9,7 +9,7 @@ module.exports = {
 			},
 		},
 		files: {
-			"assets/js/dist/yoast-premium-site-wide-analysis-420.js": [ "assets/js/src/site-wide-analysis.js" ],
+			"assets/js/dist/yoast-premium-site-wide-analysis-430.js": [ "assets/js/src/site-wide-analysis.js" ],
 			"assets/js/dist/yoast-premium-social-preview-410.js": [ "assets/js/src/yoast-premium-social-preview.js" ],
 			"assets/js/dist/wpseo-premium-contact-support-370.js": [ "assets/js/src/contact-support.js" ],
 			"assets/js/dist/wp-seo-premium-admin-redirects-370.js": [ "assets/js/src/admin-redirects.js" ],
@@ -17,8 +17,8 @@ module.exports = {
 			"assets/js/dist/wp-seo-premium-quickedit-notification-352.js": [ "assets/js/src/quickedit-notification.js" ],
 			"assets/js/dist/wpseo-premium-yoast-overlay-350.js": [ "assets/js/src/yoast-overlay.js" ],
 			"assets/js/dist/wp-seo-premium-redirect-notifications-352.js": [ "assets/js/src/redirect-notifications.js" ],
-			"assets/js/dist/wp-seo-premium-metabox-420.js": [ "assets/js/src/metabox.js" ],
 			"assets/js/dist/yoast-premium-gsc-440.js": [ "assets/js/src/google-search-console.js" ],
+			"assets/js/dist/wp-seo-premium-metabox-430.js": [ "assets/js/src/metabox.js" ],
 		},
 	},
 	"release-es6": {