+++ premium/changelog.md
@@ -1,3 +1,17 @@
+### 4.3.2: February 16th, 2017
+* Fixes a bug where CSS & JavaScript would be loaded with incorrect version numbers.
+
+### 4.3.1: February 16th, 2017
+* Fixes a bug where a slow query would be run on every page of the admin.
+
+### 4.3.0: February 14th, 2017
+* Adds link suggestions and insights for Spanish.
+* Adds variable that indicates the recalculation is running already to prevent double executed recalculations.
+* Adds a warning when there are posts or pages that need to be analyzed for link suggestions.
+* Improves the site wide calculation of the prominent words for the link suggestions. 
+* Fixes a bug where the rendering of the columns in the post overview was incorrect. 
+* Includes every change in Yoast SEO core 4.3.0, see the [core changelog](https://wordpress.org/plugins/wordpress-seo/changelog/).
+
 ### 4.2.1: February 2nd, 2017
 * Includes every change in Yoast SEO core 4.2.1, see the [core changelog](https://wordpress.org/plugins/wordpress-seo/changelog/).